backend/pyproject.toml
[project]
name = "community-learning-api"
version = "0.1.0"
description = "Community Learning Platform backend API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "sqlalchemy>=2.0.30,<3.0.0",
  "alembic>=1.13.0,<2.0.0",
  "asyncpg>=0.30.0,<1.0.0",
  "aiosqlite>=0.20.0,<1.0.0",
  "pydantic-settings>=2.3.0,<3.0.0",
  "python-multipart>=0.0.9,<1.0.0",
  "python-jose[cryptography]>=3.3.0,<4.0.0",
  "passlib[bcrypt]>=1.7.4,<2.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<9.0.0",
  "pytest-asyncio>=0.23.0,<1.0.0",
  "pytest-cov>=5.0.0,<6.0.0",
  "httpx>=0.27.0,<1.0.0",
  "ruff>=0.6.0,<1.0.0",
  "mypy>=1.10.0,<2.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
